Arreglado el boton eliminar

+++ js/carrito.js
@@ -4,7 +4,7 @@ let botonCompra = document.getElementById("carritoCompra");
 let botonVaciar = document.getElementById("vaciar-carrito");
 let contadorCarrito = document.getElementById("contador-carrito");
 let precioTotal = document.getElementById("precio-total");
-let fragmento =  document.createDocumentFragment();
+let fragmento = document.createDocumentFragment();
 let carrito = [];
 function cargarEventos() {
     document.addEventListener("DOMContentLoaded", () => {
@@ -12,26 +12,42 @@ function cargarEventos() {
         actualizarCarrito();
     })
     botonCompra.addEventListener("click", () => {
-        carrito.length = 0;
-        actualizarCarrito();
-        carrito = JSON.parse(localStorage.getItem("carrito"));
-        localStorage.clear("carrito");
-        Swal.fire(
-            'Compra Exitosa',
-            '',
-            'success'
-        )
+        if(carrito.length === 0){
+            Swal.fire(
+                'El carrito esta vacio!',
+                '',
+                'warning'
+            )
+        }else{
+            carrito.length = 0;
+            actualizarCarrito();
+            carrito = JSON.parse(localStorage.getItem("carrito"));
+            localStorage.clear("carrito");
+            Swal.fire(
+                'Compra Exitosa',
+                '',
+                'success'
+            )
+        }
     });
     botonVaciar.addEventListener("click", () => {
-        carrito.length = 0;
-        actualizarCarrito();
-        carrito = JSON.parse(localStorage.getItem("carrito"));
-        localStorage.clear("carrito");
-        Swal.fire(
-            'Vaciaste el carrito',
-            '',
-            'error'
-        )
+        if(carrito.length === 0){
+            Swal.fire(
+                'El carrito esta vacio!!',
+                '',
+                'warning'
+            )
+        }else{
+            carrito.length = 0;
+            actualizarCarrito();
+            carrito = JSON.parse(localStorage.getItem("carrito"));
+            localStorage.clear("carrito");
+            Swal.fire(
+                'Vaciaste el carrito',
+                '',
+                'error'
+            )
+        }
     });
     contenedor.addEventListener("click", agregarAlCarrito);
 };
@@ -40,20 +56,20 @@ function agregarAlCarrito(e) {
     if (e.target.classList.contains("agregar-al-carro")) {
         const prodSelec = e.target.parentElement.parentElement;
         const nombre = prodSelec.querySelector('h4').textContent;
-		Toastify({
-			text: `${nombre}`,
-			duration: 3000,
-			style: {
-				background: 'linear-gradient(to right, #00b09b, #96c93d)'
-			}
-		}).showToast();
+        Toastify({
+            text: `${nombre}`,
+            duration: 3000,
+            style: {
+                background: 'linear-gradient(to right, #00b09b, #96c93d)'
+            }
+        }).showToast();
         dataProduc(prodSelec);
     }
 };
 function dataProduc(product) {
     const infoProd = {
         nombre: product.querySelector("h4").textContent,
-        precio: Number(product.querySelector(".price").innerText.replace("$","")),
+        precio: Number(product.querySelector(".price").innerText.replace("$", "")),
         id: product.querySelector("button").id,
         cantidad: 1,
         total: 0
@@ -88,15 +104,15 @@ const actualizarCarrito = () => {
         <p>$${precio}</p>
         <p>Cantidad: <span id="cantidad">${cantidad}</span></p>
         <p>Total: ${total}</p>
-        <button onclick = "eliminarDelCarrito(${id})" id="${id}" class="btn btn-outline-danger eliminar-carro">Elminar <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-trash3-fill" viewBox="0 0 16 16">
+        <button id="${id}" class="btn btn-outline-danger eliminar-carro">Elminar <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-trash3-fill" viewBox="0 0 16 16">
         <path d="M11 1.5v1h3.5a.5.5 0 0 1 0 1h-.538l-.853 10.66A2 2 0 0 1 11.115 16h-6.23a2 2 0 0 1-1.994-1.84L2.038 3.5H1.5a.5.5 0 0 1 0-1H5v-1A1.5 1.5 0 0 1 6.5 0h3A1.5 1.5 0 0 1 11 1.5Zm-5 0v1h4v-1a.5.5 0 0 0-.5-.5h-3a.5.5 0 0 0-.5.5ZM4.5 5.029l.5 8.5a.5.5 0 1 0 .998-.06l-.5-8.5a.5.5 0 1 0-.998.06Zm6.53-.528a.5.5 0 0 0-.528.47l-.5 8.5a.5.5 0 0 0 .998.058l.5-8.5a.5.5 0 0 0-.47-.528ZM8 4.5a.5.5 0 0 0-.5.5v8.5a.5.5 0 0 0 1 0V5a.5.5 0 0 0-.5-.5Z"/>
         </svg></button>
         `
         contenedorCarrito.append(div);
         localStorage.setItem("carrito", JSON.stringify(carrito));
         let boton = document.getElementById(id);
-        boton.addEventListener("click", () => {
-            eliminarDelCarrito();
+        boton.addEventListener("click", (e) => {
+            eliminarDelCarrito(e, id);
             Toastify({
                 text: `${nombre}`,
                 duration: 3000,
@@ -117,7 +133,18 @@ const actualizarCarrito = () => {
         actualizarCarrito(); // Iterar sobre el carrito y mostrar su HTML
     }
 } */
-const eliminarDelCarrito = (prodId) => {
+/* const eliminarDelCarrito = (prodId) => {
     carrito = carrito.filter((prod) => prod.id !== prodId);
-	actualizarCarrito();
+    actualizarCarrito();
+}; */
+/* 
+onclick = "eliminarDelCarrito(${id})"
+*/
+const eliminarDelCarrito = (e, prodId) => {
+    let botonclick = e.target;
+    botonclick.parentElement.remove()
+    const item = carrito.find((prod) => prod.id === prodId);
+    const indice = carrito.indexOf(item);
+    carrito.splice(indice, 1);
+    actualizarCarrito();
 };
